requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
cloudinary==1.44.1
confusable-homoglyphs==3.3.1
cryptography==46.0.3
dj-database-url==3.1.0
Django==6.0.1
django-cors-headers==4.9.0
django-encrypted-model-fields==0.6.5
django-filter==25.2
django-registration==5.2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.14
gunicorn==24.1.1
inflection==0.5.1
Markdown==3.10.1
markdown-it-py==4.0.0
mdurl==0.1.2
packaging==26.0
psycopg2-binary==2.9.11
pycparser==3.0
Pygments==2.19.2
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
rich==14.2.0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
